% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phecode_mapping.R
\name{phecode_mapping}
\alias{phecode_mapping}
\title{Map Phecodes to ICD codes}
\usage{
phecode_mapping(
  input,
  output,
  date = "ICD_DATE",
  date_format = "\%Y-\%m-\%d",
  type = "ICD_TYPE",
  code = "ICD_CODE",
  ID = "GRID"
)
}
\arguments{
\item{input}{file path to the input file}

\item{output}{file path to the desired output}

\item{date}{name of column with entry date}

\item{date_format}{format of the date column (e.g., "\%Y-\%m-\%d")}

\item{type}{name of column that denotes ICD code type, as "ICD9CM" and "ICD10CM"}

\item{code}{name of column with ICD codes}

\item{ID}{name of ID column}
}
\description{
This function takes input from large-scale ICD code data and outputs
a file with the corresponding phecodes. The output file is in long
format, where each row is a combination of ID and phecode, the first
and last date that phecode was recorded for that ID, the total number
of times the phecode was recorded for that ID, and whether that ID is
considered to use this healthcare system as their "medical home" (having
at least five visits over a period of three years). The defaults are
designed for the input from the Vanderbilt BioVU system, but all input
columns may be defined as arguments.
}
